Added EntityNotFoundException to MovieUser controller.

+++ Application/MovieUsers/Commands/DeleteMovieUserById/DeleteMovieUserByIdCommandHandler.cs
@@ -1,4 +1,6 @@
-﻿using Infrastructure.Interfaces;
+﻿using Application.Exceptions;
+using Domain.Models;
+using Infrastructure.Interfaces;
 using MediatR;
 
 namespace Application.MovieUsers.Commands.DeleteMovieUserById
@@ -13,6 +15,12 @@ namespace Application.MovieUsers.Commands.DeleteMovieUserById
 
         public async Task<bool> Handle(DeleteMovieUserByIdCommand request, CancellationToken cancellationToken)
         {
+            UserMovie userMovie = await _repository.GetMovieUserByIdAsync(request.UserMovieId) ?? throw new EntityNotFoundException("MovUserMovieieUser", request.UserMovieId);
+            if (userMovie == null)
+            {
+                return false;
+            }
+
             await _repository.DeleteByIdAsync(request.UserMovieId);
             return true;
         }
